runpod/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sentence-transformers==3.0.1
google-generativeai==0.7.2
numpy==1.26.4
pydantic==2.7.4
pydantic-settings==2.3.4
runpod==1.7.0